# MemoryKnowledge/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge Service API
  description: |
    Standalone knowledge service for Code-Graph + LLM-Wiki.
    Provides endpoints for managing and querying code repositories
    and wiki knowledge bases.

    ## Multi-tenancy (001)

    Every endpoint requires the tenant identity `service_id` in the
    **`x-tdai-service-id` request header** (`^[A-Za-z0-9_-]+$`, missing/malformed → 400).
    This is the same value the kernel uses for store/VDB routing, so tenant identity is
    unified end-to-end. All lookups are scoped by `service_id`; an id-only endpoint
    targeting another tenant's resource returns 404. `service_id` is never a request-body
    field and is not echoed in responses.

    ## Asset ID Model

    Assets have globally unique, immutable IDs:

    | Asset Type  | Field           | Prefix   | Example        |
    |-------------|-----------------|----------|----------------|
    | LLM-Wiki    | `wiki_id`       | `wiki-`  | `wiki-9c1f2b`  |
    | Code-Graph  | `code_graph_id` | `cg-`    | `cg-7e3d10`    |

    ## Response Envelope

    All responses use `ApiResponseEnvelope`:
    - `code: 0` means success; non-zero means business error
    - `data` contains the payload on success, `null` on error
    - `message` provides human-readable status

  version: "1.0.0"
  contact:
    name: Knowledge Service
servers:
  - url: http://localhost:8421/v3
    description: Local development server
...
paths:
...
  /code-graph/set-credential:
    post:
      tags: [Code-Graph]
      summary: Change a code graph's Git credential (owner only, graph must be idle).
      security:
        - GitServiceBearer: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code_graph_id, user_id, credential_id]
              properties:
                code_graph_id:
                  type: string
                user_id:
                  type: string
                  description: Authenticated user identity asserted by Panel.
                credential_id:
                  type: string
                  nullable: true
                share_with_team:
                  type: boolean
                  description: Must be true when attaching a credential.
      responses:
        '200':
          description: Updated CodeGraphDetail in the success envelope.
        '400':
          description: Invalid credential or missing sharing confirmation.
        '401':
          description: Service authentication required.
        '403':
          description: Caller does not own this graph or credential scope does not match.
        '404':
          description: Graph or credential not found.
        '409':
          description: Code graph is busy.

  # ═══════════════════════════════════════════════
  # LLM-Wiki (15 endpoints)
  # ═══════════════════════════════════════════════
...
  /code-graph/create:
    post:
      tags: [Code-Graph]
      summary: Register repo + async clone & index
      description: |
        Generates code_graph_id (cg-{8char}), creates DB record + returns
        immediately (status=pending). Background queue runs clone + indexing.
        Idempotent: same (service_id, team_id, repo_url, branch) → returns existing (200).
      operationId: createCodeGraph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGraphCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CodeGraphDetail'
        '200':
          description: Idempotent hit
        default:
          $ref: '#/components/responses/ErrorResponse'
  /code-graph/list:
    post:
      tags: [Code-Graph]
      summary: List code graphs for a team
      operationId: listCodeGraph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGraphListRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CodeGraphListData'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /code-graph/get:
    post:
      tags: [Code-Graph]
      summary: Get code graph metadata by ID
      operationId: getCodeGraph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGraphGetRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CodeGraphDetail'
        '404':
          description: Code graph not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /code-graph/sync:
    post:
      tags: [Code-Graph]
      summary: Re-sync (git fetch + re-index)
      operationId: syncCodeGraph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGraphSyncRequest'
      responses:
        '202':
          description: Accepted
        '404':
          description: Code graph not found
        default:
          $ref: '#/components/responses/ErrorResponse'
  /code-graph/delete:
    post:
      tags: [Code-Graph]
      summary: Batch delete code graphs
      description: |
        Deletes code graphs in any status, including `pending`/`processing`
        (008): a delete never returns 409. In-flight clone/index is signalled to
        abort at its next checkpoint; the row is hard-deleted and instance
        pool / disk resources are released. Cross-memory ids report as "not found".
      operationId: deleteCodeGraph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGraphDeleteRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponseEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BatchDeleteResult'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    GitServiceBearer:
      type: http
      scheme: bearer
      description: Knowledge service key. Required even in legacy open mode for Git credentials.
  parameters:
    ServiceIdHeader:
      name: x-tdai-service-id
      in: header
      required: true
      description: >-
        Tenant identity `service_id` (001 multi-tenancy). Required on every endpoint;
        same value the kernel uses for store/VDB routing. Missing/malformed → 400.
      schema:
        type: string
        pattern: "^[A-Za-z0-9_-]+$"
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseEnvelope'
  schemas:
...
    ApiResponseEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: 0 = success, non-zero = business error
        message:
          type: string
        request_id:
          type: string
        data:
          nullable: true
    IdFields:
      type: object
      description: >-
        Common body id fields. Note: the tenant identity `service_id` is NOT a body
        field — it is sent via the `x-tdai-service-id` request header (001; unified
        with the kernel routing key). See the ServiceIdHeader parameter.
      properties:
        team_id:
          type: string
          description: Team ID (required for create/list)
        user_id:
          type: string
        agent_id:
          type: string
        task_id:
          type: string

    # ── Wiki Schemas ──
...
    CodeGraphStats:
      type: object
      properties:
        files:
          type: integer
        nodes:
          type: integer
        edges:
          type: integer
    CodeGraphDetail:
      type: object
      properties:
        code_graph_id:
          type: string
        team_id:
          type: string
        repo_name:
          type: string
        repo_url:
          type: string
        credential_id:
          type: string
          nullable: true
        branch:
          type: string
        commit_hash:
          type: string
          nullable: true
        service_url:
          type: string
          nullable: true
          description: Tools self-discovery base URL (includes API prefix, e.g. http://host:8421/v3)
        summary:
          type: string
          nullable: true
          maxLength: 256
        status:
          type: string
          enum: [pending, processing, ready, failed]
        sync_error:
          type: string
          nullable: true
        version:
          type: string
        owner_user_id:
          type: string
          nullable: true
        stats:
          allOf:
            - $ref: '#/components/schemas/CodeGraphStats'
          nullable: true
        last_sync_at:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
    CodeGraphCreateRequest:
      allOf:
        - $ref: '#/components/schemas/IdFields'
        - type: object
          required: [team_id, repo_url]
          properties:
            repo_url:
              type: string
              description: HTTPS or SSH URL without embedded HTTPS credentials.
            credential_id:
              type: string
              description: Owner's credential in the same service and team. HTTPS matches the hostname; SSH identities work across servers.
            share_with_team:
              type: boolean
              description: Must be true when binding a credential; the code index is shared with the team.
            branch:
              type: string
              default: main
            repo_name:
              type: string
    CodeGraphListRequest:
      allOf:
        - $ref: '#/components/schemas/IdFields'
        - type: object
          required: [team_id]
          properties:
            status:
              type: string
              enum: [pending, processing, ready, failed]
            limit:
              type: integer
              default: 20
            offset:
              type: integer
              default: 0
    CodeGraphListData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CodeGraphDetail'
        total:
          type: integer
    CodeGraphGetRequest:
      type: object
      required: [code_graph_id]
      properties:
        code_graph_id:
          type: string
    CodeGraphSyncRequest:
      type: object
      required: [code_graph_id]
      properties:
        code_graph_id:
          type: string
    CodeGraphDeleteRequest:
      type: object
      required: [code_graph_ids]
      properties:
        code_graph_ids:
          type: array
          items:
            type: string
          maxItems: 100
...
    BatchDeleteResult:
      type: object
      properties:
        deleted_ids:
          type: array
          items:
            type: string
        failed:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              reason:
                type: string